Moved gem function
